backend/requirements-cpu.txt
# Dépendances de base
fastapi==0.109.1
uvicorn==0.27.0
python-multipart>=0.0.7
python-jose>=3.3.0
passlib>=1.7.4
pydantic==1.10.13
google-cloud-storage==2.5.0

# Dépendances CPU
torch==2.0.0
torchvision==0.15.1

# Bibliothèques de traitement 3D et d'image
opencv-python>=4.9.0
trimesh>=4.0.5
numpy-stl>=2.17.1
python-dotenv==1.0.1
aiofiles>=0.7.0
pymongo==3.12.0
motor==2.5.1
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.5
requests==2.28.2
pillow>=9.5.0
numpy>=2.2.0
pandas>=2.2.0
scikit-learn>=1.4.0
aiohttp==3.7.4
firebase-admin==6.2.0
boto3==1.26.137
prometheus-client==0.17.1
python-json-logger==2.0.7

# Dépendances pour Blocky
reportlab==4.0.4
colormath==3.0.0
psutil>=5.8.0
tqdm==4.65.0

# Dépendances pour le traitement 3D
scipy>=1.12.0
matplotlib>=3.7.1
scikit-image>=0.21.0
bcrypt>=3.2.0
alembic>=1.7.0
asyncpg>=0.24.0
pytest>=6.2.0
pytest-asyncio>=0.15.0
black>=21.7b0
isort>=5.9.0
flake8>=3.9.0 